Update RenderTarget interface: use separate real and virtual sizes. Add touch/mouse coordinates translation from screen space to virtual screen space. Add GameLayer.

+++ core/src/com/github/alexeybond/gdx_commons/drawing/ProjectionMode.java
@@ -6,5 +6,5 @@ import com.badlogic.gdx.math.Matrix4;
  *
  */
 public interface ProjectionMode {
-    void setup(Matrix4 projectionMatrix, int targetWidth, int targetHeight);
+    void setup(Matrix4 projectionMatrix, float targetWidth, float targetHeight);
 }
